fix(repo): silence prisma source-map warnings and debug-port clashes
Prisma's generated client ships a runtime file referencing a sourcemap it
doesn't include, so every service's build spammed a "Failed to parse
source map" warning. auth-service already had a fix for this inlined in
its own webpack.config.js; extracted it to tools/webpack-ignore-prisma-
warnings.js and applied it to the other 11 Prisma-backed services so the
whole serve:all output is quiet.

Also sets serve.options.port: 0 in nx.json's targetDefaults so each
service's Node debugger binds a random free port instead of all of them
fighting over the default 9229 -- previously every "Starting inspector on
localhost:9229 failed: address already in use" line was harmless noise
from running multiple services at once, not an actual service-port
conflict.
