+++ nx.json
@@ -13,7 +13,10 @@
   "targetDefaults": {
     "build": { "dependsOn": ["^build"], "inputs": ["production", "^production"], "cache": true },
     "test": { "inputs": ["default", "^production"], "cache": true },
-    "lint": { "cache": true }
+    "lint": { "cache": true },
+    "serve": {
+      "options": { "port": 0 }
+    }
   },
   "defaultBase": "main"
 }
